+++ kmd/tests/test_completefiles.py
@@ -288,16 +288,16 @@ class DirectoryCompletionHookTests(JailSetup):
                                        'funny\\ dir/foo.')
 
     def test_dir_completion_hook(self):
-        # Even if a hook is installed, it never receives a
+        # If a hook is installed, it receives a
         # quoted directory name.
         called = []
         def hook(text):
             called.append((text,))
-            return text
+            return text.replace('\\', '')
 
         completer.directory_rewrite_hook = None
         completer.directory_completion_hook = hook
         self.assertEqual(self.complete('funny\\ dir/f'),
                                        'funny\\ dir/foo.')
-        self.assertEqual(called, [('funny dir/',)])
+        self.assertEqual(called, [('funny\\ dir/',)])